Refactor import paths for ImageCard components
